--- moja-druzyna/Controllers/HomeController.cs
@@ -1,42 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using moja_druzyna.Data;
-using moja_druzyna.Models;
-using System.Diagnostics;
-
-namespace moja_druzyna.Controllers
-{
-    public class HomeController : Controller
-    {
-        private readonly ApplicationDbContext _dbContext;
-        private readonly ILogger<HomeController> _logger;
-
-        public HomeController(ApplicationDbContext dbContext, ILogger<HomeController> logger)
-        {
-            _dbContext = dbContext;
-            _logger = logger;
-        }
-
-        public IActionResult Index()
-        {
-            if (!User.Identity.IsAuthenticated)
-                return Redirect("/Identity/Account/Login");
-
-            return View();
-        }
-
-        public IActionResult Privacy()
-        {
-            if (!User.Identity.IsAuthenticated)
-                return Redirect("/Identity/Account/Login");
-
-            return View();
-        }
-
-        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-        public IActionResult Error()
-        {
-            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
-        }
-    }
-}
