+++ moja-druzyna/Startup.cs
@@ -41,7 +41,7 @@ namespace moja_druzyna
             }
             else
             {
-                app.UseExceptionHandler("/Home/Error");
+                app.UseExceptionHandler("/Profile/Error");
                 // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
                 app.UseHsts();
             }
@@ -57,7 +57,7 @@ namespace moja_druzyna
             {
                 endpoints.MapControllerRoute(
                     name: "default",
-                    pattern: "{controller=Home}/{action=Index}/{id?}");
+                    pattern: "{controller=Profile}/{action=PersonalData}/{id?}");
                 endpoints.MapRazorPages();
             });
         }
