+++ moja-druzyna/Views/Profile/GdprConsents.cshtml
@@ -1,0 +1,8 @@
+﻿@{
+    ViewData["Title"] = "zgody RODO";
+}
+
+<div class="text-center">
+    <h1 class="display-4">Moje zgody RODO</h1>
+    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
+</div>
