+++ moja-druzyna/Views/Profile/PersonalData.cshtml
@@ -1,0 +1,8 @@
+﻿@{
+    ViewData["Title"] = "dane";
+}
+
+<div class="text-center">
+    <h1 class="display-4">Moje dane</h1>
+    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
+</div>
