+++ moja-druzyna/Views/Profile/Ranks.cshtml
@@ -1,0 +1,8 @@
+﻿@{
+    ViewData["Title"] = "stopnie";
+}
+
+<div class="text-center">
+    <h1 class="display-4">Moje stopnie</h1>
+    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
+</div>
