+++ moja-druzyna/Views/Profile/Roles.cshtml
@@ -1,0 +1,8 @@
+﻿@{
+    ViewData["Title"] = "role";
+}
+
+<div class="text-center">
+    <h1 class="display-4">Moje role</h1>
+    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
+</div>
