+++ moja-druzyna/Views/Profile/ServiceHistory.cshtml
@@ -1,0 +1,8 @@
+﻿@{
+    ViewData["Title"] = "historia służby";
+}
+
+<div class="text-center">
+    <h1 class="display-4">Moja historia służby</h1>
+    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
+</div>
