+++ moja-druzyna/Views/Shared/_LoginPartial.cshtml
@@ -1,6 +1,6 @@
 ﻿<div style="display: flex;">
     <a class="nav-link text-dark" asp-area="Identity" asp-page="/Account/Manage/Index" title="Manage">@User.Identity.Name</a>
-    <form class="form-inline" asp-area="Identity" asp-page="/Account/Logout" asp-route-returnUrl="@Url.Action("Index", "Home", new { area = "" })">
+    <form class="form-inline" asp-area="Identity" asp-page="/Account/Logout" asp-route-returnUrl="@Url.Action("Index", "Profile", new { area = "" })">
         <button type="submit" class="nav-link btn btn-link text-dark">Logout</button>
     </form>
 </div>
