+++ moja-druzyna/Views/Shared/_NoSidebarLayout.cshtml
@@ -19,7 +19,7 @@
             </nav>
         </header>
         <div class="center-section-container">
-            <a id="go-back" class="nav-link text-dark" asp-controller="Home" asp-action="Index">
+            <a id="go-back" class="nav-link text-dark" asp-controller="Profile" asp-action="Index">
                 <img src="~/icons/navigation/nav-left.svg" style="position: absolute; width: 40px; height: 40px; left: 0px; top: 0px; " />
             </a>
             <div class="container">
